all: Add EU_433 band definition
