+++ pkg/band/eu_433.go
@@ -1,4 +1,4 @@
-// Copyright © 2019 The Things Network Foundation, The Things Industries B.V.
+// Copyright © 2021 The Things Network Foundation, The Things Industries B.V.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -14,21 +14,17 @@
 
 package band
 
-import (
-	"go.thethings.network/lorawan-stack/v3/pkg/ttnpb"
+import "go.thethings.network/lorawan-stack/v3/pkg/ttnpb"
+
+const (
+	// EU_433 is the ID of the European 433Mhz band
+	EU_433 = "EU_433"
+
+	eu433BeaconFrequency = 434655000
 )
 
-//revive:disable:var-naming
-
-var eu_433 Band
-
-// EU_433 is the ID of the European 433Mhz band
-const EU_433 = "EU_433"
-
-//revive:enable:var-naming
-
-func init() {
-	defaultChannels := []Channel{
+var (
+	eu433DefaultChannels = []Channel{
 		{
 			Frequency:   433175000,
 			MaxDataRate: ttnpb.DATA_RATE_5,
@@ -42,9 +38,8 @@ func init() {
 			MaxDataRate: ttnpb.DATA_RATE_5,
 		},
 	}
-	const beaconFrequency = 434655000
 
-	downlinkDRTable := [8][6]ttnpb.DataRateIndex{
+	eu433DownlinkDRTable = [8][6]ttnpb.DataRateIndex{
 		{0, 0, 0, 0, 0, 0},
 		{1, 0, 0, 0, 0, 0},
 		{2, 1, 0, 0, 0, 0},
@@ -54,95 +49,4 @@ func init() {
 		{6, 5, 4, 3, 2, 1},
 		{7, 6, 5, 4, 3, 2},
 	}
-
-	eu_433 = Band{
-		ID: EU_433,
-
-		EnableADR: true,
-
-		MaxUplinkChannels: 16,
-		UplinkChannels:    defaultChannels,
-
-		MaxDownlinkChannels: 16,
-		DownlinkChannels:    defaultChannels,
-
-		// See ETSI EN 300.220-2 V3.1.1 (2017-02)
-		SubBands: []SubBandParameters{
-			{
-				MinFrequency: 433175000,
-				MaxFrequency: 434665000,
-				DutyCycle:    0.1,
-				MaxEIRP:      10.0 + eirpDelta,
-			},
-		},
-
-		DataRates: map[ttnpb.DataRateIndex]DataRate{
-			ttnpb.DATA_RATE_0: makeLoRaDataRate(12, 125000, makeConstMaxMACPayloadSizeFunc(59)),
-			ttnpb.DATA_RATE_1: makeLoRaDataRate(11, 125000, makeConstMaxMACPayloadSizeFunc(59)),
-			ttnpb.DATA_RATE_2: makeLoRaDataRate(10, 125000, makeConstMaxMACPayloadSizeFunc(59)),
-			ttnpb.DATA_RATE_3: makeLoRaDataRate(9, 125000, makeConstMaxMACPayloadSizeFunc(123)),
-			ttnpb.DATA_RATE_4: makeLoRaDataRate(8, 125000, makeConstMaxMACPayloadSizeFunc(230)),
-			ttnpb.DATA_RATE_5: makeLoRaDataRate(7, 125000, makeConstMaxMACPayloadSizeFunc(230)),
-			ttnpb.DATA_RATE_6: makeLoRaDataRate(7, 250000, makeConstMaxMACPayloadSizeFunc(250)),
-			ttnpb.DATA_RATE_7: makeFSKDataRate(50000, makeConstMaxMACPayloadSizeFunc(230)),
-		},
-		MaxADRDataRateIndex: ttnpb.DATA_RATE_5,
-
-		ReceiveDelay1:        defaultReceiveDelay1,
-		ReceiveDelay2:        defaultReceiveDelay2,
-		JoinAcceptDelay1:     defaultJoinAcceptDelay1,
-		JoinAcceptDelay2:     defaultJoinAcceptDelay2,
-		MaxFCntGap:           defaultMaxFCntGap,
-		ADRAckLimit:          defaultADRAckLimit,
-		ADRAckDelay:          defaultADRAckDelay,
-		MinRetransmitTimeout: defaultRetransmitTimeout - defaultRetransmitTimeoutMargin,
-		MaxRetransmitTimeout: defaultRetransmitTimeout + defaultRetransmitTimeoutMargin,
-
-		DefaultMaxEIRP: 12.15,
-		TxOffset: []float32{
-			0,
-			-2,
-			-4,
-			-6,
-			-8,
-			-10,
-		},
-
-		Rx1Channel: channelIndexIdentity,
-		Rx1DataRate: func(idx ttnpb.DataRateIndex, offset ttnpb.DataRateOffset, _ bool) (ttnpb.DataRateIndex, error) {
-			if idx > ttnpb.DATA_RATE_7 {
-				return 0, errDataRateIndexTooHigh.WithAttributes("max", 7)
-			}
-			if offset > 5 {
-				return 0, errDataRateOffsetTooHigh.WithAttributes("max", 5)
-			}
-			return downlinkDRTable[idx][offset], nil
-		},
-
-		GenerateChMasks: generateChMask16,
-		ParseChMask:     parseChMask16,
-
-		LoRaCodingRate: "4/5",
-
-		FreqMultiplier:   100,
-		ImplementsCFList: true,
-		CFListType:       ttnpb.CFListType_FREQUENCIES,
-
-		DefaultRx2Parameters: Rx2Parameters{0, 434665000},
-
-		Beacon: Beacon{
-			DataRateIndex:    ttnpb.DATA_RATE_3,
-			CodingRate:       "4/5",
-			ComputeFrequency: func(_ float64) uint64 { return beaconFrequency },
-		},
-		PingSlotFrequency: uint64Ptr(beaconFrequency),
-
-		regionalParameters1_v1_0:       bandIdentity,
-		regionalParameters1_v1_0_1:     bandIdentity,
-		regionalParameters1_v1_0_2:     bandIdentity,
-		regionalParameters1_v1_0_2RevB: bandIdentity,
-		regionalParameters1_v1_0_3RevA: bandIdentity,
-		regionalParameters1_v1_1RevA:   bandIdentity,
-	}
-	All[EU_433] = eu_433
-}
+)
